# laravel-sqlite-api-swagger/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Students API",
        "description": "API for managing students",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/players": {
            "get": {
                "tags": [
                    "Players"
                ],
                "summary": "Get all players",
                "operationId": "ed70adc58cad8f33cd0ab89dfeea7f82",
                "responses": {
                    "200": {
                        "description": "List of all players",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Player"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Players"
                ],
                "summary": "Create a new player",
                "operationId": "b146404ea10ef33cb0db469e0e2cc666",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PlayerRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Player created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Player"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/players/{id}": {
            "get": {
                "tags": [
                    "Players"
                ],
                "summary": "Get a player by ID",
                "operationId": "769c947129f093883d2a7d9cc81a42d4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Player ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Player details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Player"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Player not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Players"
                ],
                "summary": "Update a player",
                "operationId": "a09374dda47f3f193ee3c5b96fe06456",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Player ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PlayerRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "player": {
                                            "$ref": "#/components/schemas/Player"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Player not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Players"
                ],
                "summary": "Delete a player",
                "operationId": "939e7e4c2ef765d87eab30b3893cfb7c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Player ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Player not found"
                    }
                }
            }
        },
        "/api/students": {
            "get": {
                "tags": [
                    "Students"
                ],
                "summary": "Get all students",
                "operationId": "944b18b2938371d3a49103c865f5aed9",
                "responses": {
                    "200": {
                        "description": "List of all students",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Student"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Students"
                ],
                "summary": "Create a new student",
                "operationId": "55c75624de9d3fe47487bf48df409809",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StudentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Student created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Student"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/students/{id}": {
            "get": {
                "tags": [
                    "Students"
                ],
                "summary": "Get a student by ID",
                "operationId": "3d03685fc4e68bbf325c181306b4a037",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Student ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Student details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Student"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Student not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Students"
                ],
                "summary": "Update a student",
                "operationId": "c29610976f0d6bce915b31aee0e85a2f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Student ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StudentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Student not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Students"
                ],
                "summary": "Delete a student",
                "operationId": "d96ee6ae75daa51d0106f3146837dc86",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Student ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Student not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Player": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "Name": {
                        "type": "string",
                        "example": "Connor McDavid"
                    },
                    "Team": {
                        "type": "string",
                        "example": "Edmonton Oilers"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "PlayerRequest": {
                "required": [
                    "Name",
                    "Team"
                ],
                "properties": {
                    "Name": {
                        "type": "string",
                        "example": "Connor McDavid"
                    },
                    "Team": {
                        "type": "string",
                        "example": "Edmonton Oilers"
                    }
                },
                "type": "object"
            },
            "Student": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "FirstName": {
                        "type": "string",
                        "example": "John"
                    },
                    "LastName": {
                        "type": "string",
                        "example": "Doe"
                    },
                    "School": {
                        "type": "string",
                        "example": "CIT"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "StudentRequest": {
                "required": [
                    "FirstName",
                    "LastName",
                    "School"
                ],
                "properties": {
                    "FirstName": {
                        "type": "string",
                        "example": "John"
                    },
                    "LastName": {
                        "type": "string",
                        "example": "Doe"
                    },
                    "School": {
                        "type": "string",
                        "example": "CST"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Players",
            "description": "Player management endpoints"
        },
        {
            "name": "Students",
            "description": "Student management endpoints"
        }
    ]
}